Run 'bill later' once daily
